added opportunity to add new device, work with custom photoes

+++ src/main/java/net/cyberdone/commutator/model/dto/cyberplant/CyberPlantDto.java
@@ -11,16 +11,16 @@ public class CyberPlantDto {
     /**
      * Channel Settings
      */
-    private Short[] channelType;
-    private Short[] channelLogic;
+    private Integer[] channelType;
+    private Integer[] channelLogic;
     private Double[] analogSensorData;
     private Double[] maintainValue;
-    private Short[] channelValue;
-    private Short[] channelControlMode;
+    private Integer[] channelValue;
+    private Integer[] channelControlMode;
     /**
      * Channel PID Settings
      */
-    private Short[] pidDirection;
+    private Integer[] pidDirection;
     private Double[] pidKp;
     private Double[] pidKi;
     private Double[] pidKd;
@@ -28,25 +28,25 @@ public class CyberPlantDto {
     /**
      * Channel Relay Hysteresis Regulator Settings
      */
-    private Short[] relayDirection;
+    private Integer[] relayDirection;
     private Double[] relayHysteresis;
     private Double[] relayK;
     private Integer[] relayDt;
     /**
      * User Settings
      */
-    private Short sound = 0;
-    private Short waterCheck = 0;
-    private Short screensaver = 0;
-    private Short screensaverType = 0;
+    private Integer sound = 0;
+    private Integer waterCheck = 0;
+    private Integer screensaver = 0;
+    private Integer screensaverType = 0;
     private Integer screensaverAfter = 0;
-    private Short screenLight = 1;
+    private Integer screenLight = 1;
 
-    private Short dateSecond;
-    private Short dateMinute;
-    private Short dateHour;
-    private Short dateDay;
-    private Short dateMonth;
+    private Integer dateSecond;
+    private Integer dateMinute;
+    private Integer dateHour;
+    private Integer dateDay;
+    private Integer dateMonth;
     private Integer dateYear;
     /**
      * WiFi Settings
@@ -57,8 +57,8 @@ public class CyberPlantDto {
      * Developer Settings
      */
     private Integer sendDataToServerEvery = 1000;
-    private Short checkSensorEvery = 1000;
-    private Short turnOff = 0;
+    private Integer checkSensorEvery = 1000;
+    private Integer turnOff = 0;
     private Integer restartCounter = 0;
     private Integer workedTimeInSeconds = 0;
     private Integer workedTimeInHours = 0;
